requirements.txt
asn1crypto==0.24.0
azure-common==1.1.18
azure-storage-blob==1.5.0
azure-storage-common==1.4.0
boto3==1.9.130
botocore==1.12.130
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
docutils==0.14
future==0.17.1
idna==2.8
ijson==2.3
jmespath==0.9.4
pycparser==2.19
pycryptodomex==3.8.1
PyJWT==1.7.1
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
snowflake-connector-python==1.7.9
urllib3==1.24.1
